+++ src/messages/layouts/MessageLayoutElement.cpp
@@ -11,8 +11,10 @@
 #include <QDebug>
 #include <QPainter>
 #include <QPainterPath>
+#include <qtransform.h>
 
 namespace {
+using namespace chatterino;
 
 const QChar RTL_EMBED(0x202B);
 
@@ -21,6 +23,82 @@ void alignRectBottomCenter(QRectF &rect, const QRectF &reference)
     QPointF newCenter(reference.center().x(),
                       reference.bottom() - (rect.height() / 2.0));
     rect.moveCenter(newCenter);
+}
+
+void applyBttvTransformation(QPainter &painter, QRectF rect,
+                             MessageElementFlags flags)
+{
+    if (flags.has(MessageElementFlag::BttvModifierFlipV))  // vertical
+    {
+        auto scale = QTransform::fromScale(1.0, -1.0);
+        auto tl = rect.topLeft();
+        painter.setTransform(
+            QTransform::fromTranslate(-tl.x(), -tl.y()) * scale *
+            QTransform::fromTranslate(tl.x(), tl.y() + rect.height()));
+        return;
+    }
+
+    if (flags.has(MessageElementFlag::BttvModifierFlipV))  // horizontal
+    {
+        auto scale = QTransform::fromScale(-1.0, 1.0);
+        auto tl = rect.topLeft();
+        painter.setTransform(
+            QTransform::fromTranslate(-tl.x(), -tl.y()) * scale *
+            QTransform::fromTranslate(tl.x() + rect.width(), tl.y()));
+    }
+
+    if (flags.has(MessageElementFlag::BttvModifierRotateRight))
+    {
+        QTransform rotate;
+        rotate.rotate(90.0);
+        auto tl = rect.center();
+        painter.setTransform(
+            QTransform::fromTranslate(-tl.x(), -tl.y()) * rotate *
+            QTransform::fromScale(rect.width() / rect.height(),
+                                  rect.height() / rect.width()) *
+            QTransform::fromTranslate(tl.x(), tl.y()));
+    }
+
+    if (flags.has(MessageElementFlag::BttvModifierRotateLeft))
+    {
+        QTransform rotate;
+        rotate.rotate(-90.0);
+        auto tl = rect.center();
+        painter.setTransform(
+            QTransform::fromTranslate(-tl.x(), -tl.y()) * rotate *
+            QTransform::fromScale(rect.width() / rect.height(),
+                                  rect.height() / rect.width()) *
+            QTransform::fromTranslate(tl.x(), tl.y()));
+    }
+}
+
+void paintImage(QPainter &painter, QRectF rect, MessageElementFlags flags,
+                const QPixmap &pixmap)
+{
+    bool anyTransform = flags.has(MessageElementFlag::BttvModifierAnyTransform);
+    bool wideEmote = flags.has(MessageElementFlag::BttvModifierWide);
+    auto prevHints = painter.renderHints();
+
+    if (anyTransform)
+    {
+        applyBttvTransformation(painter, rect, flags);
+    }
+    if (wideEmote)
+    {
+        painter.setRenderHint(QPainter::SmoothPixmapTransform);
+    }
+
+    // fourtf: make it use qreal values
+    painter.drawPixmap(QRectF(rect), pixmap, QRectF());
+
+    if (anyTransform)
+    {
+        painter.resetTransform();
+    }
+    if (wideEmote)
+    {
+        painter.setRenderHints(prevHints);
+    }
 }
 
 }  // namespace
@@ -123,7 +201,7 @@ void ImageLayoutElement::addCopyTextToString(QString &str, uint32_t from,
         dynamic_cast<EmoteElement *>(&this->getCreator());
     if (emoteElement)
     {
-        str += emoteElement->getEmote()->getCopyString();
+        str += emoteElement->getCopyString();
         str = TwitchEmotes::cleanUpEmoteCode(str);
         if (this->hasTrailingSpace())
         {
@@ -148,8 +226,8 @@ void ImageLayoutElement::paint(QPainter &painter,
     auto pixmap = this->image_->pixmapOrLoad();
     if (pixmap && !this->image_->animated())
     {
-        // fourtf: make it use qreal values
-        painter.drawPixmap(QRectF(this->getRect()), *pixmap, QRectF());
+        paintImage(painter, this->getRect(), this->getCreator().getFlags(),
+                   *pixmap);
     }
 }
 
@@ -166,7 +244,8 @@ void ImageLayoutElement::paintAnimated(QPainter &painter, int yOffset)
         {
             auto rect = this->getRect();
             rect.moveTop(rect.y() + yOffset);
-            painter.drawPixmap(QRectF(rect), *pixmap, QRectF());
+
+            paintImage(painter, rect, this->getCreator().getFlags(), *pixmap);
         }
     }
 }
@@ -260,7 +339,8 @@ void LayeredImageLayoutElement::paint(QPainter &painter,
             QRectF destRect(0, 0, size.width(), size.height());
             alignRectBottomCenter(destRect, fullRect);
 
-            painter.drawPixmap(destRect, *pixmap, QRectF());
+            paintImage(painter, destRect, this->getCreator().getFlags(),
+                       *pixmap);
         }
     }
 }
@@ -292,7 +372,8 @@ void LayeredImageLayoutElement::paintAnimated(QPainter &painter, int yOffset)
                 QRectF destRect(0, 0, size.width(), size.height());
                 alignRectBottomCenter(destRect, fullRect);
 
-                painter.drawPixmap(destRect, *pixmap, QRectF());
+                paintImage(painter, destRect, this->getCreator().getFlags(),
+                           *pixmap);
                 animatedFlag = true;
             }
         }