+++ src/messages/layouts/MessageLayoutElement.hpp
@@ -9,9 +9,11 @@
 #include <QPoint>
 #include <QRect>
 #include <QString>
+#include <QTransform>
 
 #include <climits>
 #include <cstdint>
+#include <memory>
 
 class QPainter;