+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Minor: Added accelerators to the right click menu for messages (#4705)
 - Minor: Improved editing hotkeys. (#4628)
 - Minor: Added `/c2-theme-autoreload` command to automatically reload a custom theme. This is useful for when you're developing your own theme. (#4718)
+- Minor: Added support for BTTV emote modifiers (`w!`, `h!`, `v!`, `z!`, `!r`, `!l`). (#4478)
 - Bugfix: Fixed an issue where Subscriptions & Announcements that contained ignored phrases would still appear if the Block option was enabled. (#4748)
 - Bugfix: Increased amount of blocked users loaded from 100 to 1,000. (#4721)
 - Bugfix: Fixed pings firing for the "Your username" highlight when not signed in. (#4698)